+++ src/app/pages/tinytimes-reports/gf-channel/hour-24-channel-details.component.ts
